requirements.txt
aiogram
langchain
langchain-openai
langchain-community
langgraph

psycopg2-binary

asyncpg

pydantic-settings
python-dotenv

pydantic

indic-transliteration

pytest
pytest-asyncio